ENH: move creation of a DB engine out of adapter code
