+++ tiled/catalog/adapter.py
@@ -7,7 +7,6 @@ import itertools as it
 import logging
 import operator
 import os
-import re
 import shutil
 import sys
 import uuid
@@ -24,7 +23,6 @@ from fastapi import HTTPException, WebSocketDisconnect
 from sqlalchemy import (
     and_,
     delete,
-    event,
     exists,
     false,
     func,
@@ -38,11 +36,9 @@ from sqlalchemy import (
     update,
 )
 from sqlalchemy.dialects.postgresql import ARRAY, JSONB, REGCONFIG, TEXT
-from sqlalchemy.engine import make_url
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, create_async_engine
+from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import aliased, selectinload
-from sqlalchemy.pool import AsyncAdaptedQueuePool
 from sqlalchemy.sql.expression import cast as sql_cast
 from sqlalchemy.sql.sqltypes import MatchType
 from starlette.status import HTTP_404_NOT_FOUND, HTTP_415_UNSUPPORTED_MEDIA_TYPE
@@ -64,6 +60,7 @@ from tiled.queries import (
     StructureFamilyQuery,
 )
 
+from ..authn_database.connection_pool import get_database_engine
 from ..mimetypes import (
     APACHE_ARROW_FILE_MIME_TYPE,
     AWKWARD_BUFFERS_MIMETYPE,
@@ -76,6 +73,7 @@ from ..mimetypes import (
 from ..query_registration import QueryTranslationRegistry
 from ..server.core import NoEntry
 from ..server.schemas import Asset, DataSource, Management, Revision
+from ..server.settings import DatabaseSettings
 from ..storage import (
     FileStorage,
     SQLStorage,
@@ -101,9 +99,6 @@ from .explain import ExplainAsyncSession
 from .utils import compute_structure_id
 
 logger = logging.getLogger(__name__)
-
-DEFAULT_ECHO = bool(int(os.getenv("TILED_ECHO_SQL", "0") or "0"))
-INDEX_PATTERN = re.compile(r"^[A-Za-z0-9_-]+$")
 
 # When data is uploaded, how is it saved?
 # TODO: Make this configurable at Catalog construction time.
@@ -168,7 +163,7 @@ class RootNode:
 class Context:
     def __init__(
         self,
-        engine: AsyncEngine,
+        database_settings: DatabaseSettings,
         writable_storage=None,
         readable_storage=None,
         adapters_by_mimetype=None,
@@ -177,8 +172,8 @@ class Context:
         storage_pool_size=None,
         storage_max_overflow=None,
     ):
-        self.engine = engine
-
+        self.engine = None
+        self.database_settings = database_settings
         self.writable_storage = []
         self.readable_storage = set()
 
@@ -222,6 +217,8 @@ class Context:
 
     def session(self):
         "Convenience method for constructing an AsyncSession context"
+        if self.engine is None:
+            raise RuntimeError("Context has not been started")
         return ExplainAsyncSession(self.engine, autoflush=False, expire_on_commit=False)
 
     async def execute(self, statement, explain=None):
@@ -232,6 +229,8 @@ class Context:
             return result
 
     async def startup(self):
+        self.engine = get_database_engine(self.database_settings)
+
         if (self.engine.dialect.name == "sqlite") and (
             self.engine.url.database == ":memory:"
             or self.engine.url.query.get("mode") == "memory"
@@ -1743,7 +1742,6 @@ def in_memory(
     specs=None,
     writable_storage=None,
     readable_storage=None,
-    echo=DEFAULT_ECHO,
     adapters_by_mimetype=None,
     top_level_access_blob=None,
     cache_settings=None,
@@ -1761,7 +1759,6 @@ def in_memory(
         writable_storage=writable_storage,
         readable_storage=readable_storage,
         init_if_not_exists=True,
-        echo=echo,
         adapters_by_mimetype=adapters_by_mimetype,
         top_level_access_blob=top_level_access_blob,
         cache_settings=cache_settings,
@@ -1776,7 +1773,6 @@ def from_uri(
     writable_storage=None,
     readable_storage=None,
     init_if_not_exists=False,
-    echo=DEFAULT_ECHO,
     adapters_by_mimetype=None,
     top_level_access_blob=None,
     mount_node: Optional[Union[str, List[str]]] = None,
@@ -1788,7 +1784,7 @@ def from_uri(
 ):
     uri = ensure_specified_sql_driver(uri)
     if init_if_not_exists:
-        # The alembic stamping can only be does synchronously.
+        # The alembic stamping can only be done synchronously.
         # The cleanest option available is to start a subprocess
         # because SQLite is allergic to threads.
         import subprocess
@@ -1806,48 +1802,49 @@ def from_uri(
         logger.info(f"Subprocess stdout: {stdout}")
         logger.info(f"Subprocess stderr: {stderr}")
 
-    parsed_url = make_url(uri)
-    if (
-        (parsed_url.get_dialect().name == "sqlite")
-        and (parsed_url.database != ":memory:")
-        and (parsed_url.query.get("mode", None) != "memory")
-    ):
-        # For file-backed SQLite databases, connection pooling offers a
-        # significant performance boost. For SQLite databases that exist
-        # only in process memory, pooling is not applicable.
-        poolclass = AsyncAdaptedQueuePool
-    elif parsed_url.get_dialect().name.startswith("postgresql"):
-        poolclass = AsyncAdaptedQueuePool  # Default for PostgreSQL
-    else:
-        poolclass = None  # defer to sqlalchemy default
+    # parsed_url = make_url(uri)
+    # if (
+    #     (parsed_url.get_dialect().name == "sqlite")
+    #     and (parsed_url.database != ":memory:")
+    #     and (parsed_url.query.get("mode", None) != "memory")
+    # ):
+    #     # For file-backed SQLite databases, connection pooling offers a
+    #     # significant performance boost. For SQLite databases that exist
+    #     # only in process memory, pooling is not applicable.
+    #     poolclass = AsyncAdaptedQueuePool
+    # elif parsed_url.get_dialect().name.startswith("postgresql"):
+    #     poolclass = AsyncAdaptedQueuePool  # Default for PostgreSQL
+    # else:
+    #     poolclass = None  # defer to sqlalchemy default
 
-    node = RootNode(metadata, specs, top_level_access_blob)
-    mount_path = (
-        [segment for segment in mount_node.split("/") if segment]
-        if isinstance(mount_node, str)
-        else mount_node
-    )
+    # # Optionally set pool size and max overflow for the catalog engine;
+    # # if not specified, use the default values from sqlalchemy.
+    # pool_kwargs = (
+    #     {"pool_size": catalog_pool_size, "max_overflow": catalog_max_overflow}
+    #     if poolclass == AsyncAdaptedQueuePool
+    #     else {}
+    # )
 
-    # Optionally set pool size and max overflow for the catalog engine;
-    # if not specified, use the default values from sqlalchemy.
-    pool_kwargs = (
-        {"pool_size": catalog_pool_size, "max_overflow": catalog_max_overflow}
-        if poolclass == AsyncAdaptedQueuePool
-        else {}
+    # pool_kwargs = {k: v for k, v in pool_kwargs.items() if v is not None}
+    # # Create the async engine with the specified parameters
+    # engine = create_async_engine(
+    #     uri,
+    #     echo=echo,
+    #     json_serializer=json_serializer,
+    #     poolclass=poolclass,
+    #     **pool_kwargs,
+    # )
+    # if engine.dialect.name == "sqlite":
+    #     event.listens_for(engine.sync_engine, "connect")(_set_sqlite_pragma)
+
+    database_settings = DatabaseSettings(
+        uri=uri,
+        pool_size=catalog_pool_size,
+        max_overflow=catalog_max_overflow,
+        pool_pre_ping=False,
     )
-    pool_kwargs = {k: v for k, v in pool_kwargs.items() if v is not None}
-    # Create the async engine with the specified parameters
-    engine = create_async_engine(
-        uri,
-        echo=echo,
-        json_serializer=json_serializer,
-        poolclass=poolclass,
-        **pool_kwargs,
-    )
-    if engine.dialect.name == "sqlite":
-        event.listens_for(engine.sync_engine, "connect")(_set_sqlite_pragma)
     context = Context(
-        engine,
+        database_settings,
         writable_storage,
         readable_storage,
         adapters_by_mimetype,
@@ -1855,16 +1852,15 @@ def from_uri(
         storage_pool_size=storage_pool_size,
         storage_max_overflow=storage_max_overflow,
     )
+    node = RootNode(metadata, specs, top_level_access_blob)
+    mount_path = (
+        [segment for segment in mount_node.split("/") if segment]
+        if isinstance(mount_node, str)
+        else mount_node
+    )
     adapter = CatalogContainerAdapter(context, node, mount_path=mount_path)
 
     return adapter
-
-
-def _set_sqlite_pragma(conn, record):
-    cursor = conn.cursor()
-    # https://docs.sqlalchemy.org/en/13/dialects/sqlite.html#foreign-key-support
-    cursor.execute("PRAGMA foreign_keys=ON")
-    cursor.close()
 
 
 def format_distinct_result(results, counts):
@@ -1883,11 +1879,6 @@ class WouldDeleteData(RuntimeError):
 
 class Collision(Conflicts):
     pass
-
-
-def json_serializer(obj):
-    "The PostgreSQL JSON serializer requires str, not bytes."
-    return safe_json_dump(obj).decode()
 
 
 def key_array_to_json(keys, value):
